Factor Domain option tags into OptionTags header
